Refactor : 예약 관리 페이지 UI 수정 #72
